move relay environment into src
