+++ src/components/PlacePicker/DataList.js
@@ -4,7 +4,7 @@ import styled from "styled-components"
 import Loading from "@kiwicom/orbit-components/lib/Loading"
 
 import ListChoice from "./ListChoice"
-import { weekendapiEnvironment } from "../../../lib/enviroment"
+import { weekendapiEnvironment } from "../../services/enviroment"
 
 const StyledResults = styled.div`
   width: 100%;
